+++ packages/language-server/src/__tests__/fixtures/html/form-field-multiple-labels/index.marko
@@ -1,8 +1,15 @@
-<label for="one">A</label>
-<label for="one">B</label>
+<label for="one">
+  A
+</label>
+<label for="one">
+  B
+</label>
 <input id="one">
 
-<label for="two">A</label>
-<label>B
+<label for="two">
+  A
+</label>
+<label>
+  B
   <input id="two">
 </label>
