+++ packages/language-server/src/__tests__/fixtures/script/bound-attr-const-dynamic-member/index.marko
@@ -1,4 +1,5 @@
-<let/a = { b: 1 }/>
+<let/a={ b: 1 }/>
 
-<let/b := a["b"]/>
+<let/b:=a["b"]/>
 //   ^?   ^?
+
