+++ packages/language-server/src/language/parseScript.ts
