+++ packages/language-server/src/plugins/marko-debug/index.ts
@@ -1,6 +1,6 @@
 import { LanguageServicePlugin } from "@volar/language-service";
 import { URI } from "vscode-uri";
-import { MarkoVirtualCode } from "../core/marko-plugin";
+import { MarkoVirtualCode } from "../../language";
 
 export const create = (): LanguageServicePlugin => {
   return {
