+++ packages/language-server/src/plugins/marko/definition/index.ts
@@ -1,15 +1,18 @@
 import { NodeType } from "@marko/language-tools";
-import { Hover } from "vscode-languageserver";
-import { MarkoVirtualCode } from "../../core/marko-plugin";
+import { LocationLink } from "vscode-languageserver";
+import { MarkoVirtualCode } from "../../../language";
+import { AttrName } from "./AttrName";
 import { OpenTagName } from "./OpenTagName";
 
-export function provideHover(
+export function provideDefinitions(
   doc: MarkoVirtualCode,
   offset: number,
-): Hover | undefined {
+): LocationLink[] | undefined {
   const node = doc.markoAst.nodeAt(offset);
 
   switch (node?.type) {
+    case NodeType.AttrName:
+      return AttrName(node, doc);
     case NodeType.OpenTagName:
       return OpenTagName(node, doc);
     default:
