+++ packages/language-server/src/plugins/marko/document-symbols.ts
