+++ packages/language-server/src/plugins/marko/hover/OpenTagName.ts
@@ -2,7 +2,7 @@ import type { Node } from "@marko/language-tools";
 import { Hover } from "vscode-languageserver";
 import getTagNameCompletion from "../util/get-tag-name-completion";
 import { START_LOCATION } from "../../../utils/constants";
-import { MarkoVirtualCode } from "../../core/marko-plugin";
+import { MarkoVirtualCode } from "../../../language";
 
 export function OpenTagName(
   node: Node.OpenTagName,
